Split the bluetooth code out from the net crates.
