+++ components/bluetooth_traits/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bluetooth_traits"
+version = "0.0.1"
+authors = ["The Servo Project Developers"]
+license = "MPL-2.0"
+publish = false
+
+[lib]
+name = "bluetooth_traits"
+path = "lib.rs"
+
+[dependencies]
+ipc-channel = "0.5"
+serde = "0.8"
+serde_derive = "0.8"
