+++ components/constellation/Cargo.toml
@@ -11,6 +11,7 @@ path = "lib.rs"
 
 [dependencies]
 backtrace = "0.2.1"
+bluetooth_traits = { path = "../bluetooth_traits" }
 canvas = {path = "../canvas"}
 canvas_traits = {path = "../canvas_traits"}
 compositing = {path = "../compositing"}
