+++ components/script/Cargo.toml
@@ -25,6 +25,7 @@ angle = {git = "https://github.com/servo/angle", branch = "servo"}
 app_units = "0.3"
 audio-video-metadata = "0.1.2"
 bitflags = "0.7"
+bluetooth_traits = {path = "../bluetooth_traits"}
 canvas_traits = {path = "../canvas_traits"}
 caseless = "0.1.0"
 cookie = {version = "0.2.5", features = ["serialize-rustc"]}
